requirements.txt
python-lambda
azure-storage
azure-functions
xlrd